v0.0.012 | Better AI and Health
Refined the blob AI so it can see you, where it will then start chasing you. if you run out of it's sight, it will go to your last known position. Finaly, it has an idle state where it goes to random points in a small radius of itself. Made and added a Player Health actor compnent and made a decrease health function, but is not currently used.