v0.1.322 | Ocean VFX and Barrier Implementation
Added the ocean VFX, waves appear at the top of the level where you can see the water moving and is not constant. Underneath where fish swims is lit as if you were actually underwater and the visibility is not full, has an edge to not seeing the oil monsters fully until they're closer in range. Also added a barrier so that the fish can't swim through the top of the level.