v0.1.325 | Basic drone AI, Timer, OilBlob AI enhancement
Implemented drones very simply. I made a drone manager blueprint in the AI folder that needs to be placed within the level. This blueprint handles the timer and spawning drones at a specific time. When the player goes into the hitbox of the manager, the timer starts, and the drones will spawn randomly within the cylender on top. The drones themselves re-use the roam code the OilBlobs use, and detect if a OilBLob is within one of it's hitboxes. From there, it forces the OilBlob to move towards the drone, getting destroyed when it overlaps the drone. Also added a check to see if anything that isn't a character is infront of both the drone and the OilBlob, if so, its velocity gets inverted. PS: I had to overwrite like everything in the AI folder, so i had to redo adding the new OilBlob mesh and animation. IDK if i did that right so if someone could check that, that would be great.