v0.3.8081 | Cel Shader and Volume Adjustments
Added in the global cel shader, can be added to any material just make sure to make an instance. Ran into an issue where cel shader breaks the caustics of the ocean floor so that'll be a later date. The volume adjustment for oil blobs is done and volume slider will be implemented next.